edge-node/pyproject.toml
[project]
name = "distributed-sensing-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "flask>=3.1.0",
]
